+++ examples/hotel_receptionist/policies/cancellation.md
@@ -1,0 +1,9 @@
+Cancellation, deposit, and no-show terms for room bookings: the cancellation window, what's charged and when, and how no-shows work.
+
+Cancellation: a confirmed booking can be cancelled free of charge up to 48 hours before check-in, with a full refund to the card on file (usually two to five business days). Cancelling inside the 48-hour window retains one room-night and refunds the rest.
+
+Deposit and payment timing: the hotel charges the full stay total to the card at the time of booking - there is no separate partial deposit or staged deposit schedule. (A card is also required at check-in to cover incidentals.) If a caller asks specifically about a "deposit", be straight with them: it's the full total up front, not a deposit-and-balance arrangement - don't invent a deposit percentage or schedule.
+
+No-show: a booking is guaranteed by the card on file. A guest who neither arrives nor cancels is a no-show; because the room was held all night for them, the card is charged for the reserved stay as guaranteed. The way to avoid a no-show charge is to cancel before the window above - so encourage callers who aren't sure to cancel in time rather than simply not turning up.
+
+Taxes and extras: room rates are quoted before tax; room tax is 12%, shown at booking, and optional extras (breakfast, valet, late checkout, and the like) are itemized on the folio. Late checkout, when available, is a flat added fee.
