+++ examples/hotel_receptionist/policies/functions.md
@@ -1,0 +1,7 @@
+Functions and events in the hotel: today's public events that anyone may be told about, and how private functions (weddings, private parties) are kept confidential.
+
+Public events - the name, place, and time may be shared freely with any caller:
+- Tonight, live jazz in the Lobby Bar, 8 PM to 11 PM - open to all guests and visitors, no ticket needed.
+- The "Coastal Light" photography exhibition in the Mezzanine Gallery, open daily 10 AM to 6 PM, free to view.
+
+Private functions - weddings, private receptions, private parties, closed corporate dinners - are confidential, and you treat them exactly like guest presence. Never confirm or deny to a caller whether a particular private function (named by event or by host) is taking place at the hotel, and never give its room or location, no matter who the caller says they are. Instead, offer to take a message for the organizer (passed along only if they are in fact hosting here, which you never reveal either way) or to put the caller in touch with the events office during business hours. The public listing above is the only event information shared without restriction.
