+++ examples/hotel_receptionist/policies/safe_deposit.md
@@ -4,6 +4,6 @@ In-room safe: every room has a small electronic safe in the closet, set with a c
 
 Safe-deposit boxes: complimentary safe-deposit boxes are held behind the front desk for anything a guest would rather not leave in the room - larger amounts of cash, jewelry, travel documents, anything valuable. Ask at the desk to open one; it's free for the length of the stay.
 
-Access: in person at the desk, by the guest on the booking, with photo ID - it's not opened for anyone else and there's no access by phone. Available whenever the desk is staffed, which is 24 hours. At check-out, stop by the desk to empty the box and hand back the key before you settle up - the box isn't cleared automatically, so don't leave without it.
+Access: in person at the desk, by the guest on the booking, with photo ID - it's not opened for anyone else and there's no access by phone. Available whenever the desk is staffed, which is 24 hours. Check-out access works the same way, including an early check-out: come to the front desk in person, as the guest on the booking and with photo ID, to empty the box and hand back the key before you settle up - the 24-hour desk means there's no wrong time, but it's still in person with ID, never by phone and never cleared for you automatically, so don't leave without doing it.
 
 Liability: valuables should go in the safe-deposit box or the in-room safe. Items properly deposited in a front-desk box are covered up to the limit posted at the desk. For valuables left loose in the room and not deposited, the hotel's liability is limited - that's standard, and it's exactly why the safe and the deposit boxes are there. Set that expectation kindly and steer the guest toward depositing anything they'd hate to lose, rather than just reciting the limitation.
